pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.6",
    "django-cors-headers>=4.8.0",
    "django-extensions>=4.1",
    "django-filter>=25.1",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "flake8>=7.3.0",
    "markdown>=3.9",
    "pillow>=11.3.0",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "stripe>=10.0.0",
    "werkzeug>=3.1.3",
    "boto3>=1.35.68",
    "django-storages>=1.14.4",
    "gunicorn>=23.0.0",
]
